+++ src/components/createStudy/SelectDate.tsx
@@ -1,5 +1,6 @@
 import { ChangeEvent } from 'react'
-import { useAppDispatch } from 'src/lib/hooks/useAppDispatch'
+import { useAppDispatch } from 'src/lib/hooks'
+
 import { changeStudyForm } from 'src/lib/store/studyFormSlice'
 import styled from 'styled-components'
