+++ src/components/createStudy/SelectMaxMemberCount.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent } from 'react'
-import { useAppDispatch } from 'src/lib/hooks/useAppDispatch'
-import { useAppSelector } from 'src/lib/hooks/useAppSelector'
+import { useAppDispatch, useAppSelector } from 'src/lib/hooks'
+
 import { changeStudyForm } from 'src/lib/store/studyFormSlice'
 import styled from 'styled-components'
