+++ src/lib/hooks/useAppDispatch.ts
@@ -1,4 +1,6 @@
 import { useDispatch } from 'react-redux'
 import { AppDispatch } from '../store'
 
-export const useAppDispatch = () => useDispatch<AppDispatch>()
+const useAppDispatch = () => useDispatch<AppDispatch>()
+
+export default useAppDispatch
