+++ src/lib/hooks/useAppSelector.ts
@@ -1,4 +1,6 @@
 import { TypedUseSelectorHook, useSelector } from 'react-redux'
 import { RootState } from '../store'
 
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector
+const useAppSelector: TypedUseSelectorHook<RootState> = useSelector
+
+export default useAppSelector
