+++ src/pages/main/index.tsx
@@ -1,13 +1,14 @@
 import { useEffect } from 'react'
 import Header from 'src/components/Header/Header'
 import { refreshTokenAPI } from 'src/lib/api/refreshTokenAPI'
-import { useAppDispatch } from 'src/lib/hooks/useAppDispatch'
+
 import { getMemberInfo } from 'src/lib/store/member/memberSlice'
 import { Banner, StudyDetailInfoModal, StudyList, StudyStatusFilter } from 'src/components'
 import SearchStudy from 'src/components/mainPage/SearchStudy'
 import PageList from 'src/components/mainPage/PageList'
 import StudyTypeFilter from 'src/components/mainPage/StudyTypeFilter'
 
+import { useAppDispatch } from 'src/lib/hooks'
 import * as S from './styles'
 import { DivisionLine } from '../createStudy/styles'
